Implemented new feature in HandshakeDialog class:
Drop-down menu provided by HandshakeDialog now populates directly from
the null-terminated array of pointers to port structures generated by
the SerialComm class. Previously, the names of ports were being parsed
from the internal array of pointers to ports, and then those names were
passed via a vector to the HandshakeDialog class.

This change allows for greater flexibility in choosing what port
information is displayed in the drop-down menu, and it eliminates
the unnecessary generation of a string vector to store names in
temporarily. Additionally, it allows us to eliminate the class
variable vector previously passed by reference to any class that
calls the getter method.

The old getter method getPorts() has been depricated, and is now
renamed to getPortNames();

The functionality remains, but will likely be removed in the future
if a use-case is not found for it.