Integrate event-driven color changing display item
As of this update, the app is able to launch a new project,
conduct handshake with the connected device, and read data
from the device. A background polling thread gets data
readings, and triggers an event in the GUI that causes a
color change in a control visible in ProjectFrame.