Refactored codebase into single Main_Project
Before this update, separate versions of the project were being tracked
concurrently inside Project_Builds. However, we realized this defeats
the point of using git.

To address this issue, I made a tag for the latest stable version for
tracking purposes, then took that latest version and created a new
Directory to contain only that single version.

As of now, all working code should be contained in the top-level
directory "Main_Project". Future changes should be merged in such a way
that this is the only source code modified. Tags should be used to track
incremental changes.