Brought latest changes from Asma to Main
In this version, the sidebar contains controls to open and load
projects, while a toolbar at the top is used for in-project control.
Sensor objects have been integrated, and additional auxiliary displays
have been added.

In terms of project structure, a new directory has been added in both
the src and include directories. This directly is named "data", and
contains the code responsible for displaying different types of data.