Updated Arduino code
In this commit, I have rewritten the code that will be used for the
microcontroller. The code initializes serial communication at 9600 baud,
then runs the loop() continuously. Each iteration, it checks for
incoming commands, and parses them if present. After parsing and
executing commands, a data frame is generated, which consists of a
string of comma separated values read from connected sensors.

COMMANDS:	- "add,<string sensorname>,<int pin#>\n"
		this command causes the arduino to begin adding readings
		from the specified pin and adding them to the dataframe.
		The reason for having the name as a parameter is because
		a vector is maintained in both the microcontroller as
		well as the host application that keeps track of the
		indexes associated with each sensor, and its
		corresponding value in the dataframe.

		- "remove,<string sensorname>\n"
		this command removes all entries related to the
		specified sensor from the microcontroller memory. The
		program no longer reads from the associated pin.

		- "adjust,<int Hz>\n"
		this command causes the dataframes not to be generated
		by the microcontroller until a specified time has been
		accumulated by the device. The time resets every time a
		dataframe is generated. The input value for Hz,
		corresponds to the number of times per second that a
		dataframe will be generated.

		- "ping\n"
		this command simply causes the microcontroller to write
		"pong\n" into the serial port output. This is used to
		confirm proper execution of the microcontroller loop,
		and parsing operations.

DATAFRAME:	The data frame is formatted as follows:

		- internal vector sensornames:
		<sensor1,sensor2,sensor3,...,sensorn>

		- internal map from sensorname to pin#:
		<sensor1:pin,sensor2:pin,sensor3:pin,...,sensorn:pin>

		- dataframe sent through output buffer:
		<reading1, reading2, reading 3, ... , reading n>

The order of the dataframe values is determined by iterating through the
internal vector of sensor names, reading from the pin associated to that
name in the internal map, and then adding it to the dataframe string to
send.

As long as the states of the sensorname vectors are matching at all
times, there will be no ambiguity over which reading belongs to which
sensor. As sensors are added or removed, the order of the vectors may
change, but that is fine, once again, as long as both vectors match,
because the dataframe is generated dynamically each time by iterating
through the vector names.