sensor character limit and stability patch
As of this update, sensor names are only permitted to be up to 20
characters long. This is enforced by AddSensorDialog

Additionally, a static set of strings has been added to SerialComm
class, so that SerialComm objects can register "in-use" ports with the
set, so that all instances can see and avoid conflicts with that port.
previously, there was no enforcement, and multiple SerialComm instances
could try to open the same port, which led to a fatal error that crashed
the application.

As of this update, SerialComm checks to see if the specified port name
is already registered before attempting to handshake. It will throw an
informative error if so, otherwise it will register the name and proceed
with the handshake. When a ProjectPanel destructs, and by association,
the SerialComm instance owned by it, the SerialComm instance will
free the port resources, and then de-register the port name with the
global set so that others can use it.